update lighttpd.conf to use mod_alias
Remove userdir and use alias.url instead. This gives flexibility to
change the urls. For example, /~fsckd isntead of /~fsckdaemon.

In preparation to making the main site proxy to gunicorn, move /media
outside of document-root for archwomen.org and use an alias to match
the url.

Put all alias in conditional block for archwomen.org. This makes them
specific to the main domain and not the other subdomains. Do not want
to see foo.archwomen.org/~fsckd.