lighttpd: 404 dotfiles
Dotfiles are already hidden in directory listings but are still
viewable if one types the URL. This rule prevents them from being
served. An exception are user web dirs; let one share whatever one
wants to. In the future, restrict this rule to /media hierarchy only.

Other domains may or may not need this rule and it should be not be
applied globally. For example, in Redmine, if one makes a wiki page
beginning with a '.' then this rule would deny access to it.

Also fix some formatting. Should have made that a separate commit.
Sorry.