move Redmine from subdomain to subdir
By moving to the main domain, Redmine is accessible over https.

Unicorn requires no changes and Redmine needs only two lines.

Unfortunately Redmine does not handle the switch gracefully. It rewrites
URLs and correctly responds to most pages. For stylesheets, images,
javascripts, and favicon.ico it replies with 404. There may be a poorly
documented option to make Redmine correctly respond to those files. Until
it is found, a second instance of lighttpd will rewrite the
problematic URLs before sending the requests to Redmine.

The reason for a second instance of lighttpd is because URL rewrites
cannot be done in a $HTTP["url"] conditional. An alternative is to install
lighttpd-svn.

Old URLs to the old redmine domain are redirected to the new subdir URLs.

Redmine now uses Arch Women's favicon.ico.